+++ internal/movies_service.go
@@ -791,3 +791,45 @@ func (m *MoviesService) GetMovieSentiments(ctx context.Context, id *string) (*st
 
 	return result, resp, nil
 }
+
+// GetMovieWatchNow Returns streaming and watch now sources for a movie in the requested country.
+//
+// API docs: https://docs.trakt.tv/reference/getmovieswatchnow
+func (m *MoviesService) GetMovieWatchNow(ctx context.Context, id *string, country *string, opts *uri.ListOptions) (map[string]*str.WatchNowSources, *str.Response, error) {
+	var url = fmt.Sprintf("movies/%s/watchnow/%s", *id, *country)
+	url, err := uri.AddQuery(url, opts)
+	if err != nil {
+		return nil, nil, err
+	}
+	req, err := m.client.NewRequest(http.MethodGet, url, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	result := map[string]*str.WatchNowSources{}
+	resp, err := m.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return result, resp, nil
+}
+
+// GetMovieJustwatchLinks Returns JustWatch links for a movie in the requested country.
+//
+// API docs: https://docs.trakt.tv/reference/getmoviesjustwatchlink
+func (m *MoviesService) GetMovieJustwatchLinks(ctx context.Context, id *string, country *string) (map[string]string, *str.Response, error) {
+	var url = fmt.Sprintf("movies/%s/watchnow/justwatch_links/%s", *id, *country)
+	req, err := m.client.NewRequest(http.MethodGet, url, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	result := map[string]string{}
+	resp, err := m.client.Do(ctx, req, &result)
+	if err != nil {
+		return nil, resp, err
+	}
+
+	return result, resp, nil
+}