+++ str/watch_now_webos_link.go
@@ -1,0 +1,12 @@
+// Package str used for structs
+package str
+
+// WatchNowWebosLink represents JSON webOS link object
+type WatchNowWebosLink struct {
+	ID     *string              `json:"id,omitempty"`
+	Params *WatchNowWebosParams `json:"params,omitempty"`
+}
+
+func (w WatchNowWebosLink) String() string {
+	return Stringify(w)
+}